requirements.txt
gym==0.26.2
h5py==3.9.0
imageio==2.33.1
ipdb==0.13.13
leap==2021.1
lightning==2.0.6
lingam==1.8.1
matplotlib==3.7.1
nerv==0.2.5
networkx==2.5
numpy==1.24.3
opencv_python==4.8.1.78
pandas==2.0.3
Pillow==10.2.0
pymunk==6.6.0
pytest==7.4.0
python_igraph==0.11.3
pytorch_lightning==2.0.6
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.11.4
setuptools==68.0.0
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.9.0
xvfbwrapper==0.2.9
